Change PlayerMask to unsigned int
This should prevent the game from crashing with more than 8 players
(or at least make progress towards this goal).

+++ src/ta3d/src/stdafx.h
@@ -58,7 +58,7 @@ namespace TA3D
 
 	// widely used domain types
 	typedef int PlayerId;
-	typedef byte PlayerMask;
+	typedef unsigned int PlayerMask;
 
 	inline PlayerMask toPlayerMask(PlayerId playerId) { return 1 << playerId; }
 }
